moved normalize/translate into price.rs and moved tests over
