+++ contracts/mint/src/test.rs
@@ -6,7 +6,7 @@ pub mod tests {
         Extern, StdError, Uint128,
     };
     use mockall_double::double;
-    use shade_protocol::mint::{HandleMsg, InitMsg, QueryAnswer, QueryMsg};
+    use shade_protocol::mint::{HandleMsg, InitMsg, QueryAnswer, QueryMsg, normalize_price, translate_price};
 
     use crate::{
         contract::{handle, init, query},
